[DEV-4469] Move tests away from wrong controllers//test_entries file
